Added separate thread for slow query, and moved fetching images
Added a separate thread for loading images from the database, which takes a long time and can block the main thread.
Moved the functionality of fetching images from the JWiki class to Controller, so that that the JWiki has less responsibility, and making the image-fetching method fetchImageFromURL more flexible, such as getting local images like imageDefault.